Move CoreData protocol implementations to separate files
