+++ EssentialDevelopper/EssentialDeveloperFramework/Feed Cache/Infrastructure/CoreDataFeedStore+ImageStore.swift
