+++ EssentialDevelopper/EssentialDeveloperFrameworkTests/Feed Cache/CoreDataFeedStoreTests.swift
@@ -39,9 +39,8 @@ class CoreDataFeedStoreTests: XCTestCase, FeedStoreSpecs, SerialFeedStore {
     }
     
     func makeSUT() -> CoreDataFeedStore {
-        let bundle = Bundle(for: CoreDataFeedStore.self)
         let storeURL = URL(fileURLWithPath: "/dev/null")
-        let sut = try! CoreDataFeedStore(bundle: bundle, storeURL: storeURL)
+        let sut = try! CoreDataFeedStore(storeURL: storeURL)
         
         trackMemoryLeaks(sut)
